Lockscreen charging info: show decimal in battery temperature
[@neobuddy89: Adapted for Android 11]

+++ packages/SettingsLib/src/com/android/settingslib/fuelgauge/BatteryStatus.java
@@ -55,11 +55,11 @@ public class BatteryStatus {
     public final int maxChargingVoltage;
     public final int maxChargingWattage;
     public final boolean present;
-    public final int temperature;
+    public final float temperature;
 
     public BatteryStatus(int status, int level, int plugged, int health,
             int maxChargingWattage, boolean present,
-	    int maxChargingCurrent, int maxChargingVoltage, int temperature) {
+	    int maxChargingCurrent, int maxChargingVoltage, float temperature) {
 
         this.status = status;
         this.level = level;
